FlaskProject/requirements.txt
openai
python-dotenv
flask
flask-cors
pyjwt
werkzeug
pyodbc